Store a 64-bit displ in Instruction, closes #121

+++ src/csharp/Intel/Iced.UnitTests/Intel/DecoderTests/MemoryTest16.cs
@@ -29,7 +29,7 @@ namespace Iced.UnitTests.Intel.DecoderTests {
 	public sealed class MemoryTest16 : DecoderTest {
 		[Theory]
 		[MemberData(nameof(Test16_DecodeMemOps_as16_Data))]
-		void Test16_DecodeMemOps(string hexBytes, Code code, Register register, Register prefixSeg, Register segReg, Register baseReg, Register indexReg, int scale, uint displ, int displSize, ConstantOffsets constantOffsets, string encodedHexBytes, DecoderOptions options) =>
+		void Test16_DecodeMemOps(string hexBytes, Code code, Register register, Register prefixSeg, Register segReg, Register baseReg, Register indexReg, int scale, ulong displ, int displSize, ConstantOffsets constantOffsets, string encodedHexBytes, DecoderOptions options) =>
 			DecodeMemOpsBase(16, hexBytes, code, register, prefixSeg, segReg, baseReg, indexReg, scale, displ, displSize, constantOffsets, encodedHexBytes, options);
 		public static IEnumerable<object[]> Test16_DecodeMemOps_as16_Data => GetMemOpsData(16);
 	}
