+++ src/csharp/Intel/Iced/Intel/MemoryOperand.cs
@@ -50,7 +50,7 @@ namespace Iced.Intel {
 		/// <summary>
 		/// Memory displacement
 		/// </summary>
-		public readonly int Displacement;
+		public readonly long Displacement;
 
 		/// <summary>
 		/// 0 (no displ), 1 (16/32/64-bit, but use 2/4/8 if it doesn't fit in a <see cref="sbyte"/>), 2 (16-bit), 4 (32-bit) or 8 (64-bit)
@@ -72,7 +72,7 @@ namespace Iced.Intel {
 		/// <param name="displSize">0 (no displ), 1 (16/32/64-bit, but use 2/4/8 if it doesn't fit in a <see cref="sbyte"/>), 2 (16-bit), 4 (32-bit) or 8 (64-bit)</param>
 		/// <param name="isBroadcast"><see langword="true"/> if it's broadcasted memory (EVEX instructions)</param>
 		/// <param name="segmentPrefix">Segment override or <see cref="Register.None"/></param>
-		public MemoryOperand(Register @base, Register index, int scale, int displacement, int displSize, bool isBroadcast, Register segmentPrefix) {
+		public MemoryOperand(Register @base, Register index, int scale, long displacement, int displSize, bool isBroadcast, Register segmentPrefix) {
 			SegmentPrefix = segmentPrefix;
 			Base = @base;
 			Index = index;
@@ -108,7 +108,7 @@ namespace Iced.Intel {
 		/// <param name="displSize">0 (no displ), 1 (16/32/64-bit, but use 2/4/8 if it doesn't fit in a <see cref="sbyte"/>), 2 (16-bit), 4 (32-bit) or 8 (64-bit)</param>
 		/// <param name="isBroadcast"><see langword="true"/> if it's broadcasted memory (EVEX instructions)</param>
 		/// <param name="segmentPrefix">Segment override or <see cref="Register.None"/></param>
-		public MemoryOperand(Register @base, int displacement, int displSize, bool isBroadcast, Register segmentPrefix) {
+		public MemoryOperand(Register @base, long displacement, int displSize, bool isBroadcast, Register segmentPrefix) {
 			SegmentPrefix = segmentPrefix;
 			Base = @base;
 			Index = Register.None;
@@ -127,7 +127,7 @@ namespace Iced.Intel {
 		/// <param name="displSize">0 (no displ), 1 (16/32/64-bit, but use 2/4/8 if it doesn't fit in a <see cref="sbyte"/>), 2 (16-bit), 4 (32-bit) or 8 (64-bit)</param>
 		/// <param name="isBroadcast"><see langword="true"/> if it's broadcasted memory (EVEX instructions)</param>
 		/// <param name="segmentPrefix">Segment override or <see cref="Register.None"/></param>
-		public MemoryOperand(Register index, int scale, int displacement, int displSize, bool isBroadcast, Register segmentPrefix) {
+		public MemoryOperand(Register index, int scale, long displacement, int displSize, bool isBroadcast, Register segmentPrefix) {
 			SegmentPrefix = segmentPrefix;
 			Base = Register.None;
 			Index = index;
@@ -144,7 +144,7 @@ namespace Iced.Intel {
 		/// <param name="displacement">Memory displacement</param>
 		/// <param name="isBroadcast"><see langword="true"/> if it's broadcasted memory (EVEX instructions)</param>
 		/// <param name="segmentPrefix">Segment override or <see cref="Register.None"/></param>
-		public MemoryOperand(Register @base, int displacement, bool isBroadcast, Register segmentPrefix) {
+		public MemoryOperand(Register @base, long displacement, bool isBroadcast, Register segmentPrefix) {
 			SegmentPrefix = segmentPrefix;
 			Base = @base;
 			Index = Register.None;
@@ -162,7 +162,7 @@ namespace Iced.Intel {
 		/// <param name="scale">Index register scale (1, 2, 4, or 8)</param>
 		/// <param name="displacement">Memory displacement</param>
 		/// <param name="displSize">0 (no displ), 1 (16/32/64-bit, but use 2/4/8 if it doesn't fit in a <see cref="sbyte"/>), 2 (16-bit), 4 (32-bit) or 8 (64-bit)</param>
-		public MemoryOperand(Register @base, Register index, int scale, int displacement, int displSize) {
+		public MemoryOperand(Register @base, Register index, int scale, long displacement, int displSize) {
 			SegmentPrefix = Register.None;
 			Base = @base;
 			Index = index;
@@ -209,7 +209,7 @@ namespace Iced.Intel {
 		/// <param name="base">Base register or <see cref="Register.None"/></param>
 		/// <param name="displacement">Memory displacement</param>
 		/// <param name="displSize">0 (no displ), 1 (16/32/64-bit, but use 2/4/8 if it doesn't fit in a <see cref="sbyte"/>), 2 (16-bit), 4 (32-bit) or 8 (64-bit)</param>
-		public MemoryOperand(Register @base, int displacement, int displSize) {
+		public MemoryOperand(Register @base, long displacement, int displSize) {
 			SegmentPrefix = Register.None;
 			Base = @base;
 			Index = Register.None;
@@ -226,7 +226,7 @@ namespace Iced.Intel {
 		/// <param name="scale">Index register scale (1, 2, 4, or 8)</param>
 		/// <param name="displacement">Memory displacement</param>
 		/// <param name="displSize">0 (no displ), 1 (16/32/64-bit, but use 2/4/8 if it doesn't fit in a <see cref="sbyte"/>), 2 (16-bit), 4 (32-bit) or 8 (64-bit)</param>
-		public MemoryOperand(Register index, int scale, int displacement, int displSize) {
+		public MemoryOperand(Register index, int scale, long displacement, int displSize) {
 			SegmentPrefix = Register.None;
 			Base = Register.None;
 			Index = index;
@@ -241,7 +241,7 @@ namespace Iced.Intel {
 		/// </summary>
 		/// <param name="base">Base register or <see cref="Register.None"/></param>
 		/// <param name="displacement">Memory displacement</param>
-		public MemoryOperand(Register @base, int displacement) {
+		public MemoryOperand(Register @base, long displacement) {
 			SegmentPrefix = Register.None;
 			Base = @base;
 			Index = Register.None;
